Doclint shouldn't warn on inapplicable conditions
overview.html and doc-files/**/*.html files cannot have
a main description or be an empty comment. At the very least,
the check for being "an empty comment" for such files cannot be
performed by checking if the files contain any block tags. Block tags
are applicable to a program element, which those files are not.