Do not warn on inline misuse
The type of a declaration (module, class or interface, constructor,
method, etc.) for which a tag is applicable, is orthogonal to the type
of the tag (inline, block, bimodal).

The code up the stack knows which type of tags it has collected. If
those tags are of type other than expected, it's a programming error.